# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentRunDetail": {
        "additionalProperties": false,
        "properties": {
          "advisor_id": {
            "title": "Advisor Id",
            "type": "string"
          },
          "advisor_preferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdvisorPreferences"
              },
              {
                "type": "null"
              }
            ]
          },
          "agent_type": {
            "$ref": "#/components/schemas/AgentType"
          },
          "client_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Request Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "full_input": {
            "$ref": "#/components/schemas/RunInputSnapshot"
          },
          "has_evaluation": {
            "title": "Has Evaluation",
            "type": "boolean"
          },
          "input_hash": {
            "title": "Input Hash",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClientSummaryOutput"
              },
              {
                "$ref": "#/components/schemas/MeetingNotesOutput"
              },
              {
                "$ref": "#/components/schemas/InvestmentReviewOutput"
              }
            ],
            "title": "Output"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "provider_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Request Id"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "token_usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenUsage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "owner_id",
          "advisor_id",
          "agent_type",
          "prompt_version",
          "model_name",
          "input_hash",
          "latency_ms",
          "has_evaluation",
          "created_at",
          "full_input",
          "output"
        ],
        "title": "AgentRunDetail",
        "type": "object"
      },
      "AgentRunPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentRunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "AgentRunPage",
        "type": "object"
      },
...
      "AgentType": {
        "enum": [
          "client_summary",
          "meeting_notes",
          "investment_review"
        ],
        "title": "AgentType",
        "type": "string"
      },
...
      "DeleteResponse": {
        "additionalProperties": false,
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "resource_id": {
            "title": "Resource Id",
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "resource_id"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
      "EvaluationProvenance": {
        "additionalProperties": false,
        "properties": {
          "agent_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Model"
          },
          "benchmark_checks_enabled": {
            "title": "Benchmark Checks Enabled",
            "type": "boolean"
          },
          "caveat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caveat"
          },
          "judge_is_distinct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Is Distinct"
          },
          "judge_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Model"
          },
          "llm_judge_enabled": {
            "title": "Llm Judge Enabled",
            "type": "boolean"
          },
          "llm_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Provider"
          },
          "rule_based_enabled": {
            "default": true,
            "title": "Rule Based Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "benchmark_checks_enabled",
          "llm_judge_enabled"
        ],
        "title": "EvaluationProvenance",
        "type": "object"
      },
      "EvaluationResult": {
        "additionalProperties": false,
        "properties": {
          "advisor_usefulness": {
            "$ref": "#/components/schemas/MetricScore"
          },
          "benchmark_expectations": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetricScore"
              },
              {
                "type": "null"
              }
            ]
          },
          "clarity": {
            "$ref": "#/components/schemas/MetricScore"
          },
          "completeness": {
            "$ref": "#/components/schemas/MetricScore"
          },
          "estimated_cost": {
            "minimum": 0.0,
            "title": "Estimated Cost",
            "type": "number"
          },
          "faithfulness": {
            "$ref": "#/components/schemas/MetricScore"
          },
          "feedback": {
            "title": "Feedback",
            "type": "string"
          },
          "format_correctness": {
            "$ref": "#/components/schemas/MetricScore"
          },
          "latency_ms": {
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "number"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationProvenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_awareness": {
            "$ref": "#/components/schemas/MetricScore"
          },
          "safety": {
            "$ref": "#/components/schemas/MetricScore"
          }
        },
        "required": [
          "faithfulness",
          "completeness",
          "risk_awareness",
          "clarity",
          "advisor_usefulness",
          "safety",
          "format_correctness",
          "latency_ms",
          "estimated_cost",
          "feedback"
        ],
        "title": "EvaluationResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetricScore": {
        "additionalProperties": false,
        "properties": {
          "feedback": {
            "minLength": 1,
            "title": "Feedback",
            "type": "string"
          },
          "score": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "score",
          "feedback"
        ],
        "title": "MetricScore",
        "type": "object"
      },
...
      "PurgeRunsResponse": {
        "additionalProperties": false,
        "properties": {
          "deleted_count": {
            "minimum": 0.0,
            "title": "Deleted Count",
            "type": "integer"
          },
          "older_than_days": {
            "minimum": 0.0,
            "title": "Older Than Days",
            "type": "integer"
          }
        },
        "required": [
          "deleted_count",
          "older_than_days"
        ],
        "title": "PurgeRunsResponse",
        "type": "object"
      },
      "RunAgentRequest": {
        "additionalProperties": false,
        "properties": {
          "advisor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisor Id"
          },
          "agent_type": {
            "$ref": "#/components/schemas/AgentType"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClientSummaryPayload"
              },
              {
                "$ref": "#/components/schemas/MeetingNotesPayload"
              },
              {
                "$ref": "#/components/schemas/InvestmentReviewPayload"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "preferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdvisorPreferences"
              },
              {
                "type": "null"
              }
            ]
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          }
        },
        "required": [
          "agent_type"
        ],
        "title": "RunAgentRequest",
        "type": "object"
      },
      "RunAgentResponse": {
        "additionalProperties": false,
        "properties": {
          "advisor_id": {
            "title": "Advisor Id",
            "type": "string"
          },
          "agent_type": {
            "$ref": "#/components/schemas/AgentType"
          },
          "client_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Request Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "latency_ms": {
            "minimum": 0.0,
            "title": "Latency Ms",
            "type": "number"
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClientSummaryOutput"
              },
              {
                "$ref": "#/components/schemas/MeetingNotesOutput"
              },
              {
                "$ref": "#/components/schemas/InvestmentReviewOutput"
              }
            ],
            "title": "Output"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "provider_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Request Id"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "token_usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenUsage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "agent_type",
          "advisor_id",
          "prompt_version",
          "output",
          "latency_ms",
          "created_at"
        ],
        "title": "RunAgentResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Typed, authenticated API for evaluating and GEPA-inspired prompt optimisation of synthetic financial-advisory LLM agents.",
    "title": "Financial Advisor Agent Optimizer",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/evaluate-run/{run_id}": {
      "post": {
        "operationId": "evaluate_run_api_v1_evaluate_run__run_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Evaluate Run",
        "tags": [
          "agent runs"
        ]
      }
    },
...
    "/api/v1/run-agent": {
      "post": {
        "operationId": "run_agent_api_v1_run_agent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAgentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Run Agent",
        "tags": [
          "agent runs"
        ]
      }
    },
    "/api/v1/runs": {
      "delete": {
        "operationId": "purge_runs_api_v1_runs_delete",
        "parameters": [
          {
            "in": "query",
            "name": "older_than_days",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Older Than Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "all_owners",
            "required": false,
            "schema": {
              "default": false,
              "title": "All Owners",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeRunsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Purge Runs",
        "tags": [
          "agent runs"
        ]
      },
      "get": {
        "operationId": "runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "agent_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AgentType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Type"
            }
          },
          {
            "in": "query",
            "name": "advisor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Advisor Id"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "in": "query",
            "name": "evaluated",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Evaluated"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Runs",
        "tags": [
          "agent runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "delete": {
        "operationId": "delete_run_api_v1_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Delete Run",
        "tags": [
          "agent runs"
        ]
      },
      "get": {
        "operationId": "get_run_detail_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Run Detail",
        "tags": [
          "agent runs"
        ]
      }
    },
...
  }
}